Joint learning added in the model
New model of Lake Como featuring the joint learning approach.
This model load additional matrices of variables (time x additional dimension). These are used after the 3 BOP inputs and the exogenous signals input.
Which matrices and which column of it to use is coded into the decision variable string in the end after the parameters for the policy.